.github/workflows: use nested virtualization on Linux for BSDs
We'd like to be able to support nested virtualization on Linux rather
than macOS.  Linux VMs are less restricted in terms of capacity; they're
faster than macOS; and they're on x86-64, like our existing images,
which is helpful since the x86-64 images are the best maintained on
Vagrant.  Be sure to install Vagrant since it's not available by
default on Ubuntu systems, and to use the libvirt backend since we're
using `/dev/kvm`.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>