tests: avoid excessive verbosity by default
We'll soon add some tests that send large amounts of data to a command.
We don't want to log a megabyte worth of data and protocol messages to
process it in 16384-byte chunks, since that's a lot of data that just
serves to clutter the terminal of the developer.  Let's make the
verbosity conditional on an environment variable and default it to
quiet.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>