tests: add some tests for data streaming
We'd like to test that our command pipe data gets sent in order.  Let's
add a test to do that that processes a bunch of random data and computes
the SHA-256 hash of it.  That way, we can be sure the data is always
sent in order and correctly, since the hash will be different if the
data is not correct.

Test in both blocking and non-blocking mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>