channel: handle out-of-order writes
We'd like to allow a client to send several outstanding write requests
at once to maximize throughput.  However, our server code intentionally
dispatches each received request into a separate task and they don't
necessarily complete in order.  This is important for throughput and
because we want to allow multiple requests to be in flight at once.

Let's add a queueing functionality for writing that allows us to block
these requests (in an async-friendly way) until their preceding
operations are complete.  We keep a track of the offset of each
selector, and when it matches our request exactly, we process that
request.  If our request has an offset in the past, then we mark it as
failing due to a conflict, and if it has one in the future, then we have
to wait (using a suitable notification method) until it's ready.

The alternative would be to always execute some set of requests in
order.  However, that requires us to either parse the requests quickly
before dealing with them, which would be slow, or to guarantee that all
requests of the same message type operate in order, which could cause
blocking problems with standard output and standard error.  We also
benefit immensely from running all of the extra code in a task because
it handles panics gracefully for us without taking down the process.
Forcing the ordering behaviour to the only code that cares about it
allows us to provide improved performance and resilience, so we do that
here.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>