proxy: reserve sufficient capacity
The `Vec::reserve` method takes the desired total capacity of the
buffer, not the amount to increment it by.  Without this change, large
chunks of data, such as 65536-byte messages, can result in a panic due
to the 65571-byte range end index being larger than the buffer.

Fix this by reserving the right amount.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>