channel: avoid double close
For the `ServerCommandChannel`, we call `as_raw_fd` the file descriptor
but don't take ownership of it from the `Child`.  As a result, we
effectively close the file descriptor twice: once in the `PipeRead` or
`PipeWrite` and once in the members of the `Child` struct.

For a long time, this happened to work, but it no longer does, and Rust
prints this message:

    fatal runtime error: IO Safety violation: owned file descriptor already closed

To solve this problem, let's use `take` to grab ownership of the file
descriptor and then turn it into the `PipeRead` or `PipeWrite`, as
appropriate.  Hoist the `file_from_command` function out into the
top level of the file and use it, since we're already doing this
correctly for `ServerClipboardChannel`, from where it comes.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>