tests: add a test for out-of-order write packets
We'd like to make sure that our out-of-order write handling is robust.
To do that, let's fill a buffer with random data, intentionally send it
in an out-of-order fashion, and be sure that the command sees it in
order, as expected.  We verify that the data was received in order by
passing it through SHA-256, which will, with very high probability,
produce a different result if the data is not completely identical.

Note that we use a CSPRNG here to make sure that each block of 2048
bytes we send is different, since sending chunks of zeros wouldn't allow
us to determine if the data were actually received in order.

Since we can't completely use all of the regular client code (which
would send the packets in order, which is the prudent method one would
normally use), expose some of the internals to the crate so we can make
use of them here.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>