Provide written offset in read and write responses
It's often helpful when parsing read or write responses to have the
offset at which the data was written.  If the user is using blocking I/O
and has sent an offset, echo it back.

We know that writing out-of-order with an already-passed offset will
fail with a conflict, and doing so with a future offset will not
complete until the offset is reached, so if we're going to return a
regular success response we know that the offset must match exactly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>