Implement a complete argument for reading
When performing a blocking read, there are two possibilities for wanted
behaviour.  The first is that the read should block until _some_ data is
read; that is, it should not return EAGAIN just because there is nothing
left to read.  The second is that it should block until _all_ data is
read; that is, it should not return until the buffer is full or EOF or
an error is reached.

The former is useful for things like file system channels, where we want
to read whatever data is there incrementally but we don't want to ignore
responses until the buffer is full.  It's also useful for interactive
uses on the command line.  The latter is useful for noninteractive uses
where we want maximum throughput and buffering is allowed.

Add an option, complete, which determines which behaviour to use and
pipe it through.  Note that it only has effect in blocking mode.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>